backend/requirements-dev.txt
# Development requirements for Retail Management System
-r requirements.txt

# Code Quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.5.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
faker>=19.0.0

# Development Tools
django-debug-toolbar>=4.1.0
django-extensions>=3.2.0
ipython>=8.14.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Database Tools
django-migrations-checker>=1.0.0
